\name{maniwarp-package}
\alias{maniwarp-package}
\alias{maniwarp}
\docType{package}
\title{
\packageTitle{maniwarp}
}
\description{
\packageDescription{maniwarp}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{maniwarp}
\packageIndices{maniwarp}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{maniwarp}

Maintainer: \packageMaintainer{maniwarp}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
